Revert "Move gasestimator into the ethapi package"
This reverts commit c5af1757927e2299f7b18f8782a521473029c260.
