+++ internal/ethapi/api.go
@@ -40,6 +40,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/eth/gasestimator"
 	"github.com/ethereum/go-ethereum/eth/tracers/logger"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/p2p"
@@ -990,10 +991,10 @@ func DoEstimateGas(ctx context.Context, b Backend, args TransactionArgs, blockNr
 		return 0, err
 	}
 	// Construct the gas estimator option from the user input
-	opts := &Options{
+	opts := &gasestimator.Options{
 		Config:     b.ChainConfig(),
 		Chain:      chainContext,
-		Header:     header,
+		Header:     blockOverrides.MakeHeader(header),
 		State:      state,
 		ErrorRatio: estimateGasErrorRatio,
 	}
@@ -1008,7 +1009,7 @@ func DoEstimateGas(ctx context.Context, b Backend, args TransactionArgs, blockNr
 	call := args.ToMessage(header.BaseFee, true, true)
 
 	// Run the gas estimation and wrap any revertals into a custom return
-	estimate, revert, err := Estimate(ctx, call, opts, gasCap, blockOverrides)
+	estimate, revert, err := gasestimator.Estimate(ctx, call, opts, gasCap)
 	if err != nil {
 		if len(revert) > 0 {
 			return 0, newRevertError(revert)
